+ Refactoring the beast...
git-svn-id: https://clam-project.org/clam/branches/development-branch@5005 1f510bfa-ff1f-0410-ab8f-dacffcba9fb9

+++ CLAM/src/Processing/Analysis/OnsetDetection/OD_AudioDecimation.cxx
@@ -94,9 +94,9 @@ namespace CLAM
 			int j=0;
 			double out = 0.0;
 
-			std::deque<float> inMemory;
+			std::deque<double> inMemory;
 
-			std::deque<float> outMemory;
+			std::deque<double> outMemory;
 
 
 			// The "difficult" part of the filtering loop has been stripmined
